requirements.txt
matplotlib
requests
Pillow
Faker
python-rucaptcha
loguru